% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aes_map.R
\name{create_shape_map}
\alias{create_shape_map}
\title{Maps shapes to categorical values}
\usage{
create_shape_map(vals, map = NULL)
}
\arguments{
\item{vals}{a vector of categorical values}

\item{map}{a map definition. By default we use pch symbol identifiers.}
}
\value{
a named vector. \code{names()} are the unique values in \code{vals}, and values
are the different shapes (pch integers)
}
\description{
Map unique leves of \code{vals} to different shapes. Only works for categorical
variables.
}
