% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscatterplot.R
\name{.fscatterplot}
\alias{.fscatterplot}
\title{Generates the final plotly plot, which will be decorate with class
information by the caller}
\usage{
.fscatterplot(xx, axes, xf, yf, zf, facet_aes, facet_nrows, marker_size,
  .color, .colors, .shape, .shapes, ..., xlabel, ylabel, zlabel,
  event_source)
}
\description{
Generates the final plotly plot, which will be decorate with class
information by the caller
}
