% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{.search_df_columns}
\alias{.search_df_columns}
\title{Fast and loose way to identify which row contains a query across a data.frame}
\usage{
.search_df_columns(x, query)
}
\arguments{
\item{x}{\code{data.frame} to search}

\item{query}{the string/element to search for}
}
\description{
Fast and loose way to identify which row contains a query across a data.frame
}
