% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enkey.R
\name{enkey}
\alias{enkey}
\title{Sets the "key" column for a data.frame}
\usage{
enkey(x, key = NULL, key_column = ".data_key.", ...)
}
\arguments{
\item{x}{the data.frame}

\item{key}{The name(s) of the columns to use as a unique key into the
dataset. If key is \code{NULL} (default), then x[\link{key_column}] is added and set
to \code{1:nrow(x)}.}
}
\value{
\code{x} with the appropriate key set.
}
\description{
This column is used to match against when users brush elements on a plot.
If multiple column names are given, then a new \code{key_column} is appended
to \code{x} which is constructed by the combination of the columns specified
in \code{key}. The name of the column that was finally used as a key is stored
\code{attr(result, "key_column")}
}
