% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboxplot.R
\name{fboxplot}
\alias{fboxplot}
\alias{fboxplot.data.frame}
\title{Create an interactive boxplot}
\usage{
fboxplot(dat, x, y, with_points = FALSE, group_aes = NULL,
  color_aes = NULL, color_map = NULL, shape_aes = NULL,
  shape_map = NULL, size_aes = NULL, size_map = NULL,
  facet_aes = NULL, facet_nrows = NULL, hover = NULL, ...,
  xlabel = NULL, ylabel = NULL, marker_size = 8, sizes = c(10,
  100), pointpos = -1.8, event_source = "A")

\method{fboxplot}{data.frame}(dat, x, y, with_points = nrow(dat) < 1000,
  group_aes = NULL, color_aes = NULL, color_map = NULL,
  shape_aes = NULL, shape_map = NULL, size_aes = NULL,
  size_map = NULL, facet_aes = NULL, facet_nrows = NULL,
  hover = NULL, ..., xlabel = NULL, ylabel = NULL, marker_size = 8,
  sizes = c(10, 100), pointpos = -1.8, event_source = "A")
}
\arguments{
\item{dat}{the data source}

\item{x}{the (categorical) covariate for the x axis}

\item{y}{the (real valued) values for the y}

\item{group_aes}{column in \code{dat} to group barplots by (not yet implemented)}
}
\description{
Look at boxplot section of the new book for guidance.
https://plotly-r.com/boxplots.html
}
