% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{knit_print.FacileViz}
\alias{knit_print.FacileViz}
\title{Ensures FacileViz objects are knit properly}
\usage{
\method{knit_print}{FacileViz}(x, ...)
}
\description{
Ensures FacileViz objects are knit properly
}
