% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{maybe_facet}
\alias{maybe_facet}
\title{Drives faceting-like behavior for (internal) plotly-generating functions.}
\usage{
maybe_facet(plotfn, plotdat, facet_aes, nrows = NULL,
  has_legend = FALSE, plot_type = NULL, ..., height = 600,
  width = height + (height/6), facet_height = 300, facet_width = 370,
  legendside = NULL, widths = NULL, heights = NULL, margin = 0.02,
  shareX = TRUE, shareY = TRUE, titleX = shareX, titleY = shareY,
  which_layout = "merge")
}
\arguments{
\item{facet_aes}{The name of a column in \code{dat} to use for faceting. If this
is a length-2 vector, then we try to do a facet_grid (undone.)}

\item{nrows}{number of rows for laying out plots in a grid-like structure.
Only used if no domain is already specified.}

\item{...}{One of the following
\itemize{
\item any number of plotly/ggplot2 objects.
\item a list of plotly/ggplot2 objects.
\item a tibble with one list-column of plotly/ggplot2 objects.
}}

\item{widths}{relative width of each column on a 0-1 scale. By default all
columns have an equal relative width.}

\item{heights}{relative height of each row on a 0-1 scale. By default all
rows have an equal relative height.}

\item{margin}{either a single value or four values (all between 0 and 1).
If four values are provided, the first is used as the left margin, the second
is used as the right margin, the third is used as the top margin, and the
fourth is used as the bottom margin.
If a single value is provided, it will be used as all four margins.}

\item{shareX}{should the x-axis be shared amongst the subplots?}

\item{shareY}{should the y-axis be shared amongst the subplots?}

\item{titleX}{should x-axis titles be retained?}

\item{titleY}{should y-axis titles be retained?}

\item{which_layout}{adopt the layout of which plot? If the default value of
"merge" is used, layout options found later in the sequence of plots will
override options found earlier in the sequence. This argument also accepts a
numeric vector specifying which plots to consider when merging.}
}
\description{
Other FacileViz functions call this function to drive faceting behavior.
This function is largely meant to be internal and is purposefully not
exported yet.
}
