% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_facets.R
\name{modify_facets}
\alias{modify_facets}
\title{Update the styling of the individual facets of a ggplot.}
\usage{
modify_facets(x, specs, ...)
}
\arguments{
\item{x}{The ggplot object. We assume this has already been faceted.}

\item{specs}{A list-of-list object that defines the parameters to update.
\code{names(specs)} are the labels of the facets to update, the values are a
list of things:
* \verb{$fill}: background color
* \verb{$border}: stroke around the ribbon
* \verb{$color}: the color of the text.
* \verb{$label}: the text of the label in the facet ribbon}
}
\value{
a \code{\link[gtable:gtable]{gtable::gtable()}} object, build by \code{\link[ggplot2:ggplot_gtable]{ggplot2::ggplot_gtable()}}.
You can plot this using via \code{\link[grid:grid.draw]{grid::grid.draw()}}
}
\description{
Change label and background/fill colors of the ribbons.
}
\details{
The nitty-gritty of this function (it's rough) was inspired from many a
google search, with particularly helpful hints from CharlotteWoolley here:
https://github.com/tidyverse/ggplot2/issues/2096#issuecomment-388545380
}
\examples{
x <- ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  facet_wrap(~ cyl)
specs <- list(
  "4" = list(
    label = "four",
    fill = "red",
    color = "white"),
  "5" = list(
    fill = "purple",
    color = "yellow"))
updated <- modify_facets(x, specs)
if (interactive()) {
  grid::grid.draw(updated)
}
}
