% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot expression of a gene/feature (this function is not facile)}
\usage{
plot_expression(
  y,
  gene,
  group = "group",
  color_by = NULL,
  wrap_by = "symbol",
  inorder = TRUE,
  ...
)
}
\arguments{
\item{y}{the DGEList or voomed object your data is in}

\item{gene}{the name of the gene identifier or symbol you want to plot.}

\item{group}{what column in the $samples (or $targets, for vm) pheno table
to use to plot split expression across the x axis}

\item{color_by}{name of column (like \code{group} param) to color your points by}

\item{wrap_by}{the column in the dataset to wrap the results by. If
\code{length(gene) > 1}, you're going to need to wrwap.}
}
\value{
a ggplot object
}
\description{
This is a conveniencde function that takes your expression object
(\code{DGEList}, or "voomed" object) and plots the expression of one gene (\code{gene})
across the entire dataset grouped by the sample covariate specified by
\code{group}. You can optionally color each point by a second sample covariate
specified by the \code{color_by} parameter.
}
\examples{
y <- mbl_load_rnaseq("mouse", dataset = "mbl")
plot_expression(y, gene = "Fxyd6", group = "group",
                color_by = "source")
plot_expression(y, gene = "Fxyd6", group = "genotype",
                color_by = "source")
}
