% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unselected}
\alias{unselected}
\title{Checks if the return value from a selectInput looks like it's unselected}
\usage{
unselected(value, ignore = c("---", "__initializing__", ""))
}
\arguments{
\item{value}{The (character) object returned from a \code{selectInput}}
}
\description{
For some reason, sometimes a selectInput returns \code{NULL} and other times
it returns \code{""}, so I'm just making this utility function to deal with that
}
\details{
NOTE: This is really a function that is used by shiny modules, but instead
of having every shiny function check if something is "unselected" and setting
it to NULL, we move that functionality in here so that internal vizualization
functions can do that jusst once
}
