% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{vizualize}
\alias{vizualize}
\title{Workhorse function to generate interactive vizualizations to FacileData}
\usage{
vizualize(x, ...)
}
\arguments{
\item{x}{A "facile" object, like a \code{FacileAnalysisResult}, for one.}
}
\value{
a \code{FacileViz} object (or subclass, like \code{FacileAnalysisViz}, which
encapsulates the components required to create an interactive vizualization
from \code{x}
}
\description{
Workhorse function to generate interactive vizualizations to FacileData
}
