% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{with_aesthetics}
\alias{with_aesthetics}
\title{Utility function used by most plotting functions that returns a data.frame
with columns mapping the observations to colors, shape, etc (adds columns
for each)}
\usage{
with_aesthetics(dat, ...)
}
\description{
Utility function used by most plotting functions that returns a data.frame
with columns mapping the observations to colors, shape, etc (adds columns
for each)
}
