% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{with_shape}
\alias{with_shape}
\alias{with_shape.data.frame}
\alias{with_shape.tbl}
\title{Augment a facile object with a shape aesthetic}
\usage{
with_shape(x, aesthetic, aes_map = NULL, ..., .default_shape = I("circle"))

\method{with_shape}{data.frame}(
  x,
  aesthetic,
  aes_map = NULL,
  out_column = ".shape_aes.",
  ...,
  .default_shape = "circle"
)

\method{with_shape}{tbl}(
  x,
  aesthetic,
  aes_map = NULL,
  out_column = ".shape_aes.",
  ...,
  .default_shape = I("circle")
)
}
\arguments{
\item{x}{an object to colorize (currently just data.frame or tbls)}

\item{aesthetic}{an "attribute" from \code{x} to map levels -> shapes.
The types of object this argument can be is specific to the class of \code{x}.
For instance, for a \code{data.frame}, \code{aesthetic} would be a character vector
that specifices the column(s) used to map observations to colors.}

\item{aes_map}{unsupported for now, no way to provide your own shapes.}

\item{...}{arguments to pass down to specific implementations}

\item{.default_shape}{the default shape used when only one level exists, this
is a \code{"circile"}.}
}
\value{
the object \code{x} with an upated \code{aes_map(x, "shape")}
}
\description{
This function augments a variety of objects with shape mappings. For now we
assume objects are only data.frames or tbls.
}
\seealso{
\code{\link[=aes_map]{aes_map()}}
}
