% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesthetics.R
\name{with_size}
\alias{with_size}
\alias{with_size.data.frame}
\alias{with_size.tbl}
\title{Augment a facile object with a size aesthetic.}
\usage{
with_size(
  x,
  aesthetic,
  aes_map = NULL,
  out_column = ".size_aes.",
  ...,
  .default_size = I(1L)
)

\method{with_size}{data.frame}(
  x,
  aesthetic,
  aes_map = NULL,
  out_column = ".size_aes.",
  ...,
  .default_size = I(1L)
)

\method{with_size}{tbl}(
  x,
  aesthetic,
  aes_map = NULL,
  out_column = ".size_aes.",
  ...,
  .default_size = I(1L)
)
}
\description{
Augment a facile object with a size aesthetic.
}
