% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aes_map.R
\name{create_color_map}
\alias{create_color_map}
\title{Maps colors to categorical values}
\usage{
create_color_map(vals, map = NULL)
}
\arguments{
\item{vals}{a vector of values to create a colormap over. Curently this is
restricted to categorical vectors (character, factor), but something smart
will happen when you provide a numeric vector in due time.}

\item{map}{a map specification. defaults to a combination of
RColorBrewer Set1 and Set2 colors}
}
\value{
a named character vector, where \code{names()} are the unique levels of
\code{vals}, and the value is the color it maps to. Colors will be recycled
if there are more levels than colors provided by \code{map}.
}
\description{
Map can be a RColorBrewer name, or a vector of colors. Colors will be
recycled if \code{length(map) <}
If \code{vals}, map can be is categoricasl
}
