% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fboxplot.R
\name{fboxplot}
\alias{fboxplot}
\title{Create an interactive boxplot}
\usage{
fboxplot(dat, x, y, color_aes = NULL, color_map = NULL,
  shape_aes = NULL, shape_map = NULL, size_aes = NULL,
  size_map = NULL, hover_aes = NULL, hover_map = NULL,
  group_aes = NULL, hover = NULL, ..., xlabel = NULL,
  ylabel = NULL, zlabel = NULL, marker_size = 8, sizes = c(10,
  100), event_source = "A")
}
\arguments{
\item{dat}{the data source}

\item{x}{the (categorical) covariate for the x axis}

\item{y}{the (real valued) values for the y}

\item{color_aes}{for the color of boxplot and points}

\item{shape_aes}{for the points}

\item{size_aes}{for the points}

\item{hover_aes}{for the points}

\item{group_aes}{group bar plots along the \code{x} categorical axis}
}
\description{
Create an interactive boxplot
}
