% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscatterplot.R
\name{fscatterplot}
\alias{fscatterplot}
\alias{fscatterplot.data.frame}
\alias{fscatterplot.tbl}
\title{Draws interactive 2 or 3d scatterplots over data}
\usage{
fscatterplot(dat, axes, with_density = FALSE, color_aes = NULL,
  color_map = NULL, shape_aes = NULL, shape_map = NULL,
  size_aes = NULL, size_map = NULL, facet_aes = NULL,
  facet_nrows = NULL, hover = NULL, webgl = FALSE,
  showlegend = TRUE, ..., flat = FALSE, xlabel = NULL,
  ylabel = NULL, zlabel = NULL, marker_size = 8, height = 600,
  width = height + (height/6), sizes = c(10, 100),
  event_source = "A")

\method{fscatterplot}{data.frame}(dat, axes, with_density = FALSE,
  color_aes = NULL, color_map = NULL, shape_aes = NULL,
  shape_map = NULL, size_aes = NULL, size_map = NULL,
  facet_aes = NULL, facet_nrows = NULL, hover = NULL,
  webgl = FALSE, showlegend = TRUE, ..., flat = FALSE,
  xlabel = NULL, ylabel = NULL, zlabel = NULL, marker_size = 8,
  height = 600, width = height + (height/6), sizes = c(10, 100),
  key = NULL, event_source = "A")

\method{fscatterplot}{tbl}(dat, axes, color_aes = NULL,
  color_map = NULL, shape_aes = NULL, shape_map = NULL,
  size_aes = NULL, size_map = NULL, facet_aes = NULL,
  facet_nrows = NULL, hover = NULL, ..., xlabel = NULL,
  ylabel = NULL, zlabel = NULL, marker_size = 8, sizes = c(10,
  100), event_source = "A")
}
\arguments{
\item{axes}{the definition of the x,y,z axes}

\item{facet_aes}{The name of a column in \code{dat} to use for faceting. If this
is a length-2 vector, then we try to do a facet_grid (undone.)}

\item{webgl}{If \code{TRUE}, the plot will be rendered as WebGL, using plotly's
\code{\link[plotly:toWebGL]{plotly::toWebGL()}} function. Defaults to \code{FALSE} (svg).}

\item{...}{One of the following
\itemize{
\item any number of plotly/ggplot2 objects.
\item a list of plotly/ggplot2 objects.
\item a tibble with one list-column of plotly/ggplot2 objects.
}}

\item{x}{a data object}
}
\value{
a plotly object
}
\description{
When plotting in three dimensions, the first dimension is on the x,
second is y, and third is z.
}
\details{
Use the \code{event_source} parameter to link selection/brushing of the plot
to a listener.

TODO: Get inspired from this marginal density plot and add that functionality
here via the \code{with_density = TRUE} parameter. When this is TRUE, faceting
should be axed.
https://plotly-r.com/arranging-views.html#fig:joint
}
\examples{
dat <- data.frame(
  a = rnorm(100), b = rnorm(100), c = rnorm(100),
  class = sample(c("g1", "g2", "g3"), 100, replace = TRUE),
  grp = sample(c("n", "o", "p", "q", "r"), 100, replace = TRUE))
dat <- head(dat, 10)
fscatterplot(dat, c("a", "b"), color_aes = "class", hover = "class")
fscatterplot(dat, c("a", "b"), color_aes = "class", hover = c("class", "c"))
fscatterplot(dat, c("a", "b"), color_aes = "class",
             facet_aes = "class", hover = c("class", "c"),
             facet_nrows = 2)
fscatterplot(dat, c("a", "b"), color_aes = "class", shape_aes = "grp",
             facet_aes = "grp", hover = c("class", "c"))

# 3d
fscatterplot(dat, c("a", "b", "c"),color_aes = "class",
             hover = c("class", "c"))
fscatterplot(dat, c("a", "b", "c"), flat = TRUE, color_aes = "class",
             hover = c("class", "c"))
fscatterplot(dat, c("a", "b", "c"), color_aes = "class", shape_aes = "grp",
             hover = c("class", "c"), webgl = TRUE)
}
