% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{plot_expression}
\alias{plot_expression}
\title{Plot expression of a gene/feature (this function is not facile)}
\usage{
plot_expression(y, gene, group = "group", color_by = NULL,
  wrap_by = "symbol", ...)
}
\arguments{
\item{y}{the DGEList or voomed object your data is in}

\item{gene}{the name of the gene identifier or symbol you want to plot.}

\item{group}{what column in the $samples (or $targets, for vm) pheno table
to use to plot split expression across the x axis}

\item{color_by}{name of column (like \code{group} param) to color your points by}
}
\value{
a ggplot object
}
\description{
This is a conveniencde function that takes your expression object
(\code{DGEList}, or "voomed" object) and plots the expression of one gene (\code{gene})
across the entire dataset grouped by the sample covariate specified by
\code{group}. You can optionally color each point by a second sample covariate
specified by the \code{color_by} parameter.
}
\examples{
y <- mbl_load_rnaseq("mouse", dataset = "mbl")
mbl_plot_expression(y, gene = "Fxyd6", group = "group",
                    color_by = "source")
mbl_plot_expression(y, gene = "Fxyd6", group = "genotype",
                    color_by = "source")
}
\seealso{
\code{\link[=mbl_tidy]{mbl_tidy()}}
}
